+++ index.ts
@@ -1,0 +1,7 @@
+import type { ICredentialType, INodeTypeDescription } from 'n8n-workflow';
+
+export { GoogleAdsConversion } from './nodes/GoogleAdsConversion/GoogleAdsConversion.node';
+export { GoogleAdsOAuth2 } from './credentials/GoogleAdsOAuth2.credentials';
+
+export const nodes: INodeTypeDescription[] = [];
+export const credentials: ICredentialType[] = [];